Refactor: CustomView 클래스를 따로 분리
