+++ app/src/main/java/com/codesquard/kotlin_drawingapp/TaskPresenter.kt
@@ -10,33 +10,12 @@ class TaskPresenter(val taskView: TaskContract.TaskView) : TaskContract.Presente
     private val plane: PlaneListener = Plane()
     private val rectangleList = ArrayList<RectangleViewModel>()
 
-    override fun onCreateView(context: Context): CustomView {
+    /*override fun onCreateView(context: Context): CustomView {
         return CustomView(context)
-    }
+    }*/
 
-    inner class CustomView(context: Context) : View(context) {
-        private val rectangle = plane.onCreateRectangle()
 
-        init {
-            rectangleList.add(rectangle)
-        }
-
-        override fun onDraw(canvas: Canvas?) {
-            super.onDraw(canvas)
-
-            val paint = Paint()
-            val alpha = rectangle.getAlpha()
-            val r = rectangle.getColor()[0]
-            val g = rectangle.getColor()[1]
-            val b = rectangle.getColor()[2]
-            paint.color = Color.argb(alpha, r, g, b)
-
-            val x = rectangle.getPoint()[0]
-            val y = rectangle.getPoint()[1]
-            canvas?.drawRect(x, y, 150f + x, 120f + y, paint)
-        }
 
     }
 
 
-}
